Adjust hash field exist condition for overwrite
To not overwrite on a table with only hash key, hash should
not exist.

But to not overwrite on a table with hash and range key, we must
make the range key unique. If we make hash and range unique, it
wont insert when hash matches and range doesnt.

This requires more investigation (how to make unique the *pair*
of hash and range key).